requirements.txt
asgiref==3.6.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
distro==1.9.0
Django==4.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
easy-thumbnails==2.8.5
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
libusb==1.0.27
mysqlclient==2.2.4
openai==0.28.0
packaging==23.2
pillow==10.2.0
pkg-about==1.1.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2023.3
pyusb==1.2.1
PyYAML==6.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.0
zope.interface==6.3
